circular import refactor
